Integrated first boot hooks in the image system, fix issues with mutable files in ro nfsroot and disabled readahead in CentOS

+++ share/common.sh
@@ -1429,9 +1429,15 @@ function hooks()
 function first_boot_hooks()
 {
     local hooks_path=$1
-    cp -p ${hooks_path}/first_boot/first_boot.service  /lib/systemd/system/
-    cp -p ${hooks_path}/first_boot/first_boot /usr/local/bin/first_boot
-    chmod 700 /usr/local/bin/first_boot
+    if [[ ! -e ${hooks_path}/first_boot ]]; then 
+        mkdir -p ${hooks_path}/first_boot
+    fi
+    if [[ ! -e /usr/local/bin/first_boot ]]; then
+        cp -p ${hooks_path}/first_boot/first_boot.service  /lib/systemd/system/
+        cp -p ${hooks_path}/first_boot/first_boot /usr/local/bin/first_boot
+        chmod 700 /usr/local/bin/first_boot
+    fi
+    replace_text /usr/local/bin/first_boot "Environment=\"HOOKS_PATH=" "Environment=\"HOOKS_PATH=${hooks_path}\""
     systemctl enable first_boot
 } 1>>$LOGFILE 2>&1
 
@@ -1445,14 +1451,22 @@ function enable_readonly_root()
         ;;
         rhel|redhat|centos)
             sed -i "s|READONLY=no|READONLY=yes|g" $prefix/etc/sysconfig/readonly-root
-            chroot $prefix /usr/bin/systemctl disable systemd-readahead-collect.service
             replace_text $prefix/etc/hosts "^::1" " "
             echo "files    /etc/aliases.db" > $prefix/etc/rwtab.d/postfix
             echo "dirs     /var/lib/postfix" >> $prefix/etc/rwtab.d/postfix
+            echo "files    /var/run/gssproxy.pid" > /etc/rwtab.d/gssproxy
+            echo "dirs     /var/lib/gssproxy" >> /etc/rwtab.d/gssproxy
+            echo "dirs     /var/lock" > /etc/rwtab.d/tmpdirs
+            echo "dirs     /var/lib/rpm" >> /etc/rwtab.d/tmpdirs
         ;;
         suse|sle[sd]|opensuse)
             error_exit "Read-only NFSROOT image not yet supported for this distribution"
-            chroot $prefix /usr/bin/systemctl disable systemd-readahead-collect.service
+            echo "files    /etc/aliases.db" > $prefix/etc/rwtab.d/postfix
+            echo "dirs     /var/lib/postfix" >> $prefix/etc/rwtab.d/postfix
+            echo "files    /var/run/gssproxy.pid" > /etc/rwtab.d/gssproxy
+            echo "dirs     /var/lib/gssproxy/rcache" >> /etc/rwtab.d/gssproxy
+            echo "dirs     /var/lock" > /etc/rwtab.d/tmpdirs
+            echo "dirs     /var/lib/rpm" >> /etc/rwtab.d/tmpdirs
         ;;
         *)
             warning_msg "This distribution is not supported."
@@ -1463,10 +1477,16 @@ function enable_readonly_root()
 function generate_rootfs()
 {
     local image=$1
+    local hooks_path=${SNOW_CONF}/boot/images/$image
     # set mount point for the rootfs
     local mount_point="/dev/shm/rootfs"
     # create a mount point
     mkdir -p ${mount_point}
+    # Run hooks:
+    hooks ${hooks_path}
+    # Setup the first boot hooks
+    replace_text /usr/local/bin/first_boot "Environment=\"HOOKS_PATH=" "Environment=\"HOOKS_PATH=${hooks_path}\""
+    systemctl enable first_boot
     # Transfer required files
     rsync -aHAXv --progress --exclude=/proc/* --exclude=/sys/* --exclude=/sNow/* --exclude=/tmp/* --exclude=/dev/* --exclude=/var/log/messages / ${mount_point}/
     # Create required directory structure
@@ -1509,10 +1529,6 @@ function generate_rootfs_nfs()
     echo "sysfs       /sys        sysfs   defaults    0 0" >> ${mount_point}/etc/fstab
     setup_networkfs ${mount_point}
     enable_readonly_root ${mount_point}
-    # Run hooks:
-    hooks ${SNOW_CONF}/boot/images/$image
-    # Setup the first boot hooks
-    first_boot_hooks ${SNOW_CONF}/boot/images/$image
     # Setup NFSROOT support for PXE
     cp -p ${SNOW_CONF}/boot/pxelinux.cfg/nfsroot ${image_pxe}
     sed -i "s|__IMAGE__|$image|g" ${image_pxe}
